feat: split trail
