+++ trailblazing/functions/getFullName.ts
@@ -1,0 +1,7 @@
+import { Person } from '../models/Person';
+
+function getFullName(person: Person) {
+  return `${person.name}`;
+}
+
+export { getFullName };
