+++ trailblazing/functions/greet.ts
@@ -1,0 +1,7 @@
+import { Person } from '../models/Person';
+
+function greet(person: Person) {
+  console.log(`Hello, ${person.name}!`);
+}
+
+export { greet };
