+++ trailblazing/models/Address.ts
@@ -1,0 +1,8 @@
+interface Address {
+  street: string;
+  city: string;
+  state: string;
+  zip: string;
+}
+
+export { Address };
