+++ trailblazing/models/Person.ts
@@ -1,0 +1,10 @@
+import { Address } from './Address';
+
+interface Person {
+  name: string;
+  age: number;
+  email?: string;
+  address?: Address;
+}
+
+export { Person };
